0053_maximum-subarray_test.go
package leetcode

import (
	"reflect"
	"testing"
)

func Test_maxSubArray(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{"", args{[]int{-2, 1, -3, 4, -1, 2, 1, -5, 4}}, 6},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := maxSubArray(tt.args.nums); got != tt.want {
				t.Errorf("maxSubArray() = %v, want %v", got, tt.want)
			}
		})
	}
}

func Test_maxSubArray_return_array(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{"", args{[]int{-2, 1, -3, 4, -1, 2, 1, -5, 4}}, []int{4, -1, 2, 1}},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := maxSubArrayReturnArray(tt.args.nums); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("maxSubArray_return_array() = %v, want %v", got, tt.want)
			}
		})
	}
}
